% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_program.R
\name{add_program}
\alias{add_program}
\title{program to add details of a code file path to the meta-table}
\usage{
add_program(number, calling_prog, meta_table_string = "meta_table",
  frame = parent.frame(), ...)
}
\arguments{
\item{number}{the number of the table or figure}

\item{calling_prog}{the file path}

\item{meta_table_string}{character string of the name of a global table that contains meta information on tables (title, population, number)}

\item{frame}{an environment win which to clean up}

\item{...}{modifications to the default values \code{\link{detach_pop}}}
}
\description{
program to add details of a code file path to the meta-table
}
\keyword{internal}
