% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_is_numeric}
\alias{all_is_numeric}
\title{Check if All Elements in Character Vector are Numeric}
\usage{
all_is_numeric(x, extras = c(".", NA))
}
\arguments{
\item{x}{a character vector}

\item{extras}{a vector of character strings to count as numeric values,
other than "".}
}
\value{
a logical value
}
\description{
Tests, without issuing warnings, whether all elements of a character
vector are legal numeric values, or optionally converts the vector to
 a numeric vector. Leading and trailing blanks in x are ignored.
}
\examples{
all_is_numeric(c('1','1.2','3'))
all_is_numeric(c('1','1.2','3a'))

}
