% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_utils.R
\name{apply_macro_dict}
\alias{apply_macro_dict}
\title{Apply DLU/CLU file to the data frame}
\usage{
apply_macro_dict(data, dlu, clu = NULL, date_format = "\%d/\%m/\%Y")
}
\arguments{
\item{data}{Data frame to be applied.}

\item{dlu}{Data frame of DLU}

\item{clu}{Data frame of CLU}

\item{date_format}{Date format to be converted, default is `\%d/\%m/\%Y`.}
}
\value{
A data.table object.
}
\description{
The variable label attribute will be applied to the data frame from the DLU file.
}
\details{
This funciton first convert the data to a \code{\link[data.table]{data.table}}.
This is to avoid the variable attributes dropped by base R functions. Then it will use
the dlu file to convert the data into corresponding variable types.
\itemize{
  \item{IntegerData}: Convert to numeric.
  \item{Real}: Convert to numeric.
  \item{Category}: If there are any non-numerica characters in the variable, no conversion
will be performed, otherwise convert to numeric.
  \item{Date}: Convert data date format with \code{\link{as.Date}}. The \code{date_format}
will be used to feed the \code{\link{as.Date}} function during the covnersion.
  \item{Text}: Convert to character.
}
After the conversion of the data, variable label attribute will be created for the variable.
See  \code{\link{var_lab}}.

After this is completed and if the clu file is provided, value label attribute will be
create for the variables listed in the clu file. See \code{\link{val_lab}}.
}
\seealso{
\code{\link{var_lab}} \code{\link{val_lab}} \code{\link{sep_dlu}}
 \code{\link[data.table]{data.table}}
}
