% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_pop.R
\name{attach_pop}
\alias{attach_pop}
\alias{detach_pop}
\alias{.eval_pop}
\title{Functions that attach and detach functions based on which populations are used for a given table number}
\usage{
attach_pop(number, ...)

detach_pop(number, ...)

.eval_pop(number, function_name, meta_table_string = "meta_table",
  popn_table_string = "popn_table", ...)
}
\arguments{
\item{number}{the number of a table or figure}

\item{...}{modifications to the default values of meta_table_string, reserved_string, popn_table_string see \code{\link{clean_up}}}

\item{meta_table_string}{character string of the name of a global table that contains meta information on tables (title, population, number)}

\item{popn_table_string}{character string of the name of a gloabl table that contains population names used in meta-table, and the equivalent names of environments}
}
\value{
invisibly returns an environment for attaching, or NULL for detaching.
}
\description{
Functions that attach and detach functions based on which populations are used for a given table number
}
\section{Functions}{
\itemize{
\item \code{attach_pop}: attaches a population

\item \code{detach_pop}: detaches a populations

\item \code{.eval_pop}: internal function
}}

\seealso{
\code{\link{attach}} \code{\link{detach}}
}
\keyword{internal}
