% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_ggplot.R
\name{build_kmplot}
\alias{build_kmplot}
\title{Combine KM-plot with risk table}
\usage{
build_kmplot(x, ...)
}
\arguments{
\item{x}{km_ggplot object}

\item{...}{other arguments not used}
}
\description{
Combine KM-plot with risk table
}
\keyword{internal}
