% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctu_initialise.R
\name{cctu_initialise}
\alias{cctu_initialise}
\title{initialise the directories for using cctu package}
\usage{
cctu_initialise(root = getwd())
}
\arguments{
\item{root}{the root directory to start in}
}
\value{
invisible return of logical indicating if the directories have been created. The directories needed are "Output", and within "Output", "Core", "Figures", "Reports".
}
\description{
initialise the directories for using cctu package
}
\seealso{
\code{\link{dir.create}}
}
