% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R
\name{clean_up}
\alias{clean_up}
\title{removes all the objects in the environment, apart from those listed in reserved and detaches the environment used to create tables or figures corresponding to the number}
\usage{
clean_up(number, envir = parent.frame(), reserved_string = "RESERVED", ...)
}
\arguments{
\item{number}{a number associated with a table of figure}

\item{envir}{an environment to clean up}

\item{reserved_string}{a character giving the name of a global variable that names gloabl objects that should generally be preserved globally when tidying up.}

\item{...}{modifications to the default values of meta_table_string, reserved_string, popn_table_string see \code{\link{clean_up}}}
}
\value{
NULL, but removes objects
}
\description{
removes all the objects in the environment, apart from those listed in reserved and detaches the environment used to create tables or figures corresponding to the number
}
