% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_missing.R
\name{dump_missing_report}
\alias{dump_missing_report}
\alias{get_missing_report}
\alias{reset_missing_report}
\title{Save/Get/Reset missingness report}
\usage{
dump_missing_report(x = "Output/variable_missing_report.csv")

get_missing_report()

reset_missing_report()
}
\arguments{
\item{x}{File path the report will be dumped to. Default is under `Output`
folder, named as `variable_missing_report.csv`.}
}
\description{
\code{dump_missing_report} can be used to save the missingness report to
a file.
\code{get_missing_report} Return the missingness report data.
\code{reset_missing_report} Reset the internal missingness report data to blank.
}
\seealso{
\code{\link{cttab}} \code{\link{report_missing}}
}
