% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tree.R
\name{find_common_prefix}
\alias{find_common_prefix}
\title{Print methods for code_tree object
 S3 class inheritance means we dont' need this at all
 it will default to print.data.frame}
\usage{
find_common_prefix(vec)
}
\description{
Print methods for code_tree object
 S3 class inheritance means we dont' need this at all
 it will default to print.data.frame
}
\keyword{internal}
