% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_digits.R
\name{format_pval}
\alias{format_pval}
\title{Format p-value}
\usage{
format_pval(pvals, sig_limit = 10^(-digits), digits = 3)
}
\arguments{
\item{pvals}{A numeric value or vector of p-values}

\item{sig_limit}{Lower bound for precision; smaller values will be shown
as < sig_limit}

\item{digits}{Number of digits past the decimal point to keep}
}
\description{
Format p-value
}
\examples{
pv <- c(-1, 0.00001, 0.0042, 0.0601, 0.1335, 0.4999, 0.51, 0.89, 0.9, 1)
format_pval(pv)

}
