% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tree.R
\name{get_code_tree}
\alias{get_code_tree}
\alias{reset_code_tree}
\title{track which files of code are called by which oher files}
\usage{
get_code_tree()

reset_code_tree(root_file = "main.R")
}
\arguments{
\item{root_file}{the name of the code file you want to use as the initial
root for code tree}
}
\value{
the data file containing two variables: parent ; child.
}
\description{
track which files of code are called by which oher files
}
\details{
each time the \code{cctu} function \code{\link{source}} is called,
an internal data frame is appended with the parent file that called
\code{\link{source}}, and the child file that was sourced.

You can reset the code_tree or examine it, but not directly edit it. It is
reset automatically if
\code{\link{cctu_initialise}} is called.
}
\section{Functions}{
\itemize{
\item \code{reset_code_tree()}: reset the internal code_tree object to have no
entries

}}
\seealso{
\code{\link{cctu_initialise}}
}
