% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_name.R
\name{get_file_name}
\alias{get_file_name}
\title{Function to get the filename and path of the most deeply nested call to source(), or whatever function in pattern}
\usage{
get_file_name(pattern = "source", ...)
}
\arguments{
\item{pattern}{the name of a function to search for}

\item{...}{arguments to supply to \code{\link{normalizePath}}}
}
\value{
a character string that is the value of \code{\link{normalizePath}} applied to the first
argument of the most deeply nested call to \code{pattern}. Or, in lay terms, if you use source()
in a nested fashion, then calling get_file_name() will give the name of the file it is called within.
}
\description{
Function to get the filename and path of the most deeply nested call to source(), or whatever function in pattern
}
\seealso{
\code{\link{write_ggplot}} \code{\link{write_table}}
}
