% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obj.R
\name{get_obj}
\alias{get_obj}
\title{Gets copies of objects from parent environments, or returns alternative if not found}
\usage{
get_obj(name, frame = parent.frame(), alt = NULL)
}
\arguments{
\item{name}{text string of an object to find}

\item{alt}{an alternative to return if the object is not found}
}
\value{
either a copy of the object \code{name} or an alternative if it does not exist
}
\description{
Gets copies of objects from parent environments, or returns alternative if not found
}
\details{
this is designed to use within functions in the library to access global objects,
but provide a route to avoid hard-coding specific variable names in
}
\examples{
rm(PATH)
get_obj("PATH")
get_obj("PATH", alt=getwd())
PATH <- "C:/MyFile"
get_obj("PATH")
get_obj("PATH", alt=getwd())
}
\keyword{internal}
