% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore_init.R
\name{gitcheck}
\alias{gitcheck}
\title{Checks for problem files to avoid saving in github}
\usage{
gitcheck(
  folder = ".",
  extensions = c("csv", "xls", "xlsx", "png", "eps", "jpg", "pdf", "doc", "docx", "zip")
)
}
\arguments{
\item{folder}{The folder to run checks within. Needs to be a git repository. Defaults to current working directory}

\item{extensions}{A character vector giving the file extensions to search for}
}
\value{
invisibly returns a list of the un-tracked and tracked problem files. Will print on terminal a report as the main objective.
}
\description{
Checks for problem files to avoid saving in github
}
