% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore_init.R
\name{gitignore_init}
\alias{gitignore_init}
\title{checks to see if a .gitignore needs to be create or edited based on the template file}
\usage{
gitignore_init(
  template = system.file("extdata/gitignore_template", package = "cctu")
)
}
\arguments{
\item{template}{the template .gitignore file.  Default is part of the package data.}
}
\value{
nothing, called for its side effect.  Gives messages out if  file is created or edited
}
\description{
checks to see if a .gitignore needs to be create or edited based on the template file
}
