% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttab_plot.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Arrange multiple plots with shared legend}
\usage{
grid_arrange_shared_legend(
  plot_list,
  ncol = 1,
  nrow = ceiling(length(plot_list)/ncol),
  position = c("bottom", "right"),
  common_leg = TRUE
)
}
\arguments{
\item{plot_list}{A list of ggplot list.}

\item{ncol}{Number of columns}

\item{nrow}{Number of rows}

\item{position}{Position of legend}

\item{common_leg}{If common legend}
}
\description{
Arrange multiple plots with shared legend
}
\examples{
dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
p1 <- qplot(carat, price, data = dsamp, colour = clarity)
p2 <- qplot(cut, price, data = dsamp, colour = clarity)
p3 <- qplot(color, price, data = dsamp, colour = clarity)
p4 <- qplot(depth, price, data = dsamp, colour = clarity)
grid_arrange_shared_legend(p1, p2, p3, p4, ncol = 4, nrow = 1)
grid_arrange_shared_legend(p1, p2, p3, p4, ncol = 2, nrow = 2)

}
\references{
https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
}
\keyword{internal}
