% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{init_cctu_env}
\alias{init_cctu_env}
\title{Initialise cctu package environment and options}
\usage{
init_cctu_env()
}
\description{
This function will reset the default options. Internal use, user shouldn't call
this function during the analysis.
}
\keyword{internal}
