% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{Check whether string, list or vector is empty}
\usage{
is_empty(x, na_empty = TRUE)
}
\arguments{
\item{x}{String, character vector, list, data.frame or numeric vector
or factor.}

\item{na_empty}{Logical, if \code{NA} is considered as empty
(default is \code{TRUE}).}
}
\value{
Logical vector.
}
\description{
This function checks whether a string or character vector,
a list or any vector (numeric, atomic) is empty or not.
}
\examples{
is_empty("test")
is_empty("")
is_empty(NA)
is_empty(NULL)

# string is not empty
is_empty(" ")

# however, this trimmed string is
is_empty(trimws(" "))

# numeric vector
x <- 1
is_empty(x)
x <- x[-1]
is_empty(x)

# check multiple elements of character vectors
is_empty(c("", "a"))

# empty data frame
d <- data.frame()
is_empty(d)

# empty list
# is_empty(list(NULL))

# NA vector
x <- rep(NA, 5)
is_empty(x)
is_empty(x, na_empty = FALSE)
}
