% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_ggplot.R
\name{km_ggplot}
\alias{km_ggplot}
\title{Create a Kaplan-Meier plot using ggplot2}
\usage{
km_ggplot(
  sfit,
  xlabs = "Time",
  ylabs = "",
  strata_labs = NULL,
  ystratalabs = deprecated(),
  timeby = NULL,
  pval = FALSE,
  p_digits = getOption("cctu_p_digits", default = 4),
  ...
)
}
\arguments{
\item{sfit}{a \code{\link[survival]{survfit}} object}

\item{xlabs}{x-axis label}

\item{ylabs}{y-axis label}

\item{strata_labs}{The strata labels. If left as NULL it defaults to \code{levels(summary(sfit)$strata)} with minor prettification.}

\item{ystratalabs}{deprecated and only for backcompatibility. use strata_labs argument.}

\item{timeby}{numeric: Default is NULL to use ggplot defaults, but allows user to specify the gaps between x-axis ticks}

\item{pval}{logical: add the pvalue to the plot?}

\item{p_digits}{integer: the number of decimal places to use for a p-value.}

\item{...}{option parameters include xlims and ylims to set the axes' ranges,
where defaults are derived from the data: both are vectors of length two giving the min and max.}
}
\value{
a list of ggplot objects is made: the top figure and a table of counts.
The object has a print and plot method that uses \code{patchwork} to glue together.
The user can access and modify the ggplot components as desired.
}
\description{
Create a Kaplan-Meier plot using ggplot2
}
\examples{
library(survival)
 fit <- survfit(Surv(time,status)~rx, data=colon)
 km_ggplot(fit)
}
\author{
Originall taken from  \url{http://statbandit.wordpress.com/2011/03/08/an-enhanced-kaplan-meier-plot/} but modified by authors of \code{cctu} package.
}
