% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{lab2val}
\alias{lab2val}
\title{Replace vector/matrix/data.frame values with corresponding value labels.}
\usage{
lab2val(x)
}
\arguments{
\item{x}{vector/matrix/data.frame}
}
\value{
Object of the same form as x but with value labels instead of values.
}
\description{
\code{lab2val} replaces vector/matrix/data.frame values with
corresponding value labels. If there are no labels for some values they are
converted to characters in most cases. If there are no labels at all for
variable it remains unchanged.
}
\examples{
data(mtcars)
mtcars <- within(mtcars, {
  var_lab(mpg) <- NULL
  val_lab(am) <- c(" automatic" = 0, " manual" = 1)
})

table(lab2val(mtcars$am))

summary(lm(mpg ~ ., data = lab2val(mtcars[, c("mpg", "am")])))
}
\references{
This is a modified version from `expss` package.
}
\seealso{
\link{val_lab}, \link{var_lab}
}
