% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_description.R
\name{library_description}
\alias{library_description}
\title{loads all the packages listed in the DESCRIPTION file}
\usage{
library_description()
}
\value{
a list of the packages
}
\description{
loads all the packages listed in the DESCRIPTION file
}
