% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-data.r
\name{merge_data}
\alias{merge_data}
\title{Merge vertically split data}
\usage{
merge_data(datalist, dlulist, clulist)
}
\arguments{
\item{datalist}{A list of dataset}

\item{dlulist}{A list of DLU data}

\item{clulist}{A list of CLU data}
}
\value{
three data.frame
}
\description{
This function can be used when the MACRO dataset is \strong{vertically} split,
must not be used to combine horizontally split data which has same variable
names. This function will rename the variable of the dataset, including
shortcode of dlu and clu files with underscore and number, and combine them
to single dataset. Dataset named \code{dt}, dlu named \code{dlu} and
\code{clu} will return to current environment. Data, DLU and CLU should be in
 same order. The merged dataset, dlu and clu can further be passed to
 \code{\link{apply_macro_dict}}.
}
\examples{
\dontrun{
 merge_data(datalist = list(dt_A, dt_B),
            dlulist = list(dlu_A, dlu_B),
            clulist = list(clu_A, clu_B))
}
}
\seealso{
\code{\link{read_data}}
\code{\link{apply_macro_dict}}
}
