% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_stat.R
\name{num_stat}
\alias{num_stat}
\alias{cat_stat}
\title{Compute some basic descriptive statistics.}
\usage{
num_stat(x, digits = 3, digits.pct = 1)

cat_stat(x, digits.pct = 1)
}
\arguments{
\item{x}{A vector or numeric, factor, character or logical values.}

\item{digits}{An integer specifying the number of significant digits to keep
for numerical results. See \code{signif_pad}.}

\item{digits.pct}{An integer specifying the number of significant digits to 
keep for percentage.}
}
\value{
A list. For numeric \code{x}, the list contains the numeric elements:
\itemize{
  \item \code{N}: the number of non-missing values
  \item \code{NMISS}: the number of missing values
  \item \code{SUM}: the sum of the non-missing values
  \item \code{MEAN}: the mean of the non-missing values
  \item \code{SD}: the standard deviation of the non-missing values
  \item \code{MIN}: the minimum of the non-missing values
  \item \code{MEDIAN}: the median of the non-missing values
  \item \code{CV}: the percent coefficient of variation of the non-missing values
  \item \code{GMEAN}: the geometric mean of the non-missing values if non-negative, or \code{NA}
  \item \code{GCV}: the percent geometric coefficient of variation of the
  non-missing values if non-negative, or \code{NA}
}
If \code{x} is categorical (i.e. factor, character or logical), the list
contains a sublist for each category, where each sublist contains the
numeric elements:
\itemize{
  \item \code{FREQ}: the frequency count
  \item \code{PCT}: the percent relative frequency, including NA in the denominator
  \item \code{PCTnoNA}: the percent relative frequency, excluding NA from the denominator
}
}
\description{
Values of type \code{factor}, \code{character} and \code{logical} are
treated as categorical. For logicals, the two categories are given the
labels `Yes` for \code{TRUE}, and `No` for \code{FALSE}.  Factor levels with
zero counts are retained.
}
\examples{
x <- exp(rnorm(100, 1, 1))
num_stat(x)

y <- factor(sample(0:1, 99, replace=TRUE), labels=c("Female", "Male"))
y[1:10] <- NA
cat_stat(y)
cat_stat(is.na(y))

}
\keyword{utilities}
