% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_dewey.R
\name{order_dewey}
\alias{order_dewey}
\title{sorts inputs that look like Dewey decimal numbering}
\usage{
order_dewey(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a numeric vector that sorts the input as per dewey decimal system, without confusing 10.2, vs 10.10
}
\description{
sorts inputs that look like Dewey decimal numbering
}
