% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_tree.R
\name{plot.code_tree}
\alias{plot.code_tree}
\title{Print/plot the code tree.}
\usage{
\method{plot}{code_tree}(x, root_file = "main.R", ...)
}
\arguments{
\item{x}{code tree object from \code{\link{get_code_tree}}.}

\item{root_file}{the name of the code file you want to use as the initial
root for code tree}

\item{...}{other parameters to be passed to \code{\link[igraph]{plot.igraph}}}
}
\description{
Print/plot the code tree.
}
\examples{
\dontrun{
plot(get_code_tree(), root_file = "main.R")
}
}
