% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_ggplot.R
\name{plot.km_ggplot}
\alias{plot.km_ggplot}
\title{plot methods for km_ggplot object}
\usage{
\method{plot}{km_ggplot}(x, ...)
}
\arguments{
\item{x}{km_ggplot object}

\item{...}{other arguments for generic methods}
}
\description{
plot methods for km_ggplot object
}
