% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propercase.R
\name{propercase}
\alias{propercase}
\title{Function to convert text to proper case}
\usage{
propercase(text)
}
\arguments{
\item{text}{a character vector}
}
\value{
a character vector converted to "proper" case, where each word
starts with a capital, and all subsequent letters are lower case
}
\description{
Function to convert text to proper case
}
