% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_missing.R
\name{report_missing}
\alias{report_missing}
\title{Generate missing report}
\usage{
report_missing(
  data,
  vars,
  select,
  row_split = NULL,
  dlu = dlu,
  subjid_string = "subjid"
)
}
\arguments{
\item{data}{A \code{data.frame} from which the variables in \code{vars}
should be taken.}

\item{select}{a named vector with as many components as row-variables. Every
element of `select` will be used to select the individuals to be analyzed
 for every row-variable. Name of the vector corresponds to the row variable,
 element is the selection.}

\item{row_split}{Variable that used for splitting table rows, rows will be
 splited using this variable. Useful for repeated measures. If \code{x} is 
a \code{formula}, \code{row_split} should be provded on the far right hand
side of the formula after \code{|}. For example, \code{treat ~ age + sex|cycle}.}

\item{dlu}{A data.frame of DLU file.}

\item{subjid_string}{A character naming the column used to identify subject,
default is \code{"subjid"}.}
}
\value{
A data frame
}
\description{
This function use the same method as described in the \code{cctab}, reporting
missingness of the variables. It includes which form is the variable from,
set as `Derived` if not from DLU file. And missing percentage with which subjects
are have missing value for that particular variable.
}
\keyword{internal}
