% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_envir.R
\name{rm_envir}
\alias{rm_envir}
\title{A function to remove all attached objects from the search path with any exceptions specified}
\usage{
rm_envir(ignore = c(".GlobalEnv", "package:", "tools:", "Autoloads"))
}
\arguments{
\item{ignore}{a character vector of regular expression search terms  within the output from \code{search()} as to which should be retained.}
}
\description{
A function to remove all attached objects from the search path with any exceptions specified
}
\details{
it detaches anything not ignored.
}
\seealso{
\code{\link{search}}
}
