% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_batch.R
\name{run_batch}
\alias{run_batch}
\title{Runs a script in batch mode directly from within R.}
\usage{
run_batch(filename, ...)
}
\arguments{
\item{filename}{character string of the name or path of the script (relative to the working directory).}

\item{...}{other arguments to pass to \code{\link{shell}}}
}
\value{
see \code{\link{shell}} which is called.
}
\description{
This function is to allow R CMD BATCH to be called interactively, without needing to start
up a command line terminal and change directories etc.  It calls the R CMD BATCH from the current working directory.
Hence its value is in producing the side-effects: whatever happens when the script is run.
Plus a log file will be created in the working directory with the same name as the script but with the suffix changed to '.Rout'.
}
\details{
The command will not run in environments where \code{\link{interactive}} is false, so as to prevent
infinite loops. Only expected to work in Windows.
}
\seealso{
\code{\link{BATCH}} and \code{\link[callr]{rcmd}}, which might already do this....

\code{\link[devtools]{check_failures}} for future ideas for a function to check the log file
for errors or warnings.
}
