% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_utils.R
\name{sep_dlu}
\alias{sep_dlu}
\title{Tidy DLU form}
\usage{
sep_dlu(x)
}
\arguments{
\item{x}{DLU data.frame}
}
\value{
A data.frame
}
\description{
Separate Visit, Form and Question into different columns.
The DLU file contains four columns:
\itemize{
  \item{ShortCode}: Variable name of the downloaded data.
  \item{Visit/Form/Question}: Contains visit, form and question. It is
separated by a slash. This function will separate this column into `Visit`,
`Form` and `Question` column in the output dataset. The `Question` column
is the unique variable name for a particular CRF form.
  \item{Description}: Description of the variable, namely variable label.
This is will be used by \code{\link{apply_macro_dict}} to create variable label.
 \item{Type}: Type of the variable, it has IntegerData, Text, Date, Real and
Category four categories. This will be used by \code{\link{apply_macro_dict}} to
convert variables to corresponding type.
}
}
\seealso{
\code{\link{apply_macro_dict}}
}
