% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_utils.R
\name{set_dlu}
\alias{set_dlu}
\alias{get_dlu}
\title{Set/get DLU data}
\usage{
set_dlu(x, clean_names = TRUE)

get_dlu()
}
\arguments{
\item{x}{DLU data}

\item{clean_names}{Conver variable name to lower case (default). See 
\code{\link{clean_names}} for details.}
}
\description{
\code{set_dlu} will set the provided DLU data to package environment, so it can be used
 for missing data report by \code{\link{cttab}}. It is user's responsibility to make
 sure values of the short code in the provided DLU data matches the variable names in
 the dataset. \code{set_dlu} will not have any effect on \code{\link{apply_macro_dict}}.
Instead, the \code{\link{apply_macro_dict}} will override the `DLU` seetings done by
\code{set_dlu}.  

\code{get_dlu} can be used to get a copy of DLU data in stored by \code{set_dlu}.
}
