% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumby.R
\name{sumby}
\alias{sumby}
\title{Produces summary statistics on a variable broken down by arm, display a graphical equivalent.}
\usage{
sumby(variable, arm, label = NULL, data = parent.frame(), total = TRUE,
  fig = TRUE, directory = "Output/Figures/")
}
\arguments{
\item{variable}{A variable of interest to be summarised. Works for several classes: numeric, integer, factor, character.}

\item{arm}{A variable to break down the summary by, typically an arm in a RCT. Or a constant variable for 1-armed study}

\item{label}{a text string to label the variable by, used to add in units or full words with spaces rather than acronyms. Defaults to NULL.}

\item{data}{optional a data frame, or envirnoment, in which to find the variable and arm values. Defaults to parent.frame()}

\item{total}{logical value to include an extra overall columns. Defaults to true}

\item{fig}{logical value on whether to print the figure. Defaults to true}
}
\value{
a data.frame containing summary statistics in character format,
ready to use with write_table(). Plus an attribute "fig" that contains a ggplot object
}
\description{
Produces summary statistics on a variable broken down by arm, display a graphical equivalent.
}
\seealso{
\code{\link{sumfig}} , \code{\link{write_table}}
}
