% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumby_ae.R
\name{sumby_ae}
\alias{sumby_ae}
\title{Function to calculate summary statistics by group and by site for AEs/ADEs}
\usage{
sumby_ae(safety_data, rand_data, varname, bysite = FALSE)
}
\arguments{
\item{safety_data}{safety dataset - must have variables subjid, description, rx, siteid}

\item{rand_data}{randomisation dataset - must have variable rx}

\item{varname}{label for ae description - e.g. "ADE Description"}

\item{bysite}{if TRUE then summaries calculated by site as well as treatment group}
}
\value{
a table with summary statistics in
}
\description{
Function to calculate summary statistics by group and by site for AEs/ADEs
}
