% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_utils.R
\name{tidy_dlu}
\alias{tidy_dlu}
\title{Tidy DLU form}
\usage{
tidy_dlu(x, clean_names = TRUE)
}
\arguments{
\item{x}{DLU data}

\item{clean_names}{Conver variable name to lower case (default). See 
\code{\link{clean_names}} for details.}
}
\value{
A data.frame with `visit/form/question` column separated into `visit`,
`form` and `question` column.
}
\description{
Separate visit, form and question into different columns.
Variable names, NOT the values, of the dlu data will be converted to lower cases.
}
\details{
The DLU file contains four columns:
\itemize{
  \item{shortcode}: Variable name of the downloaded data.
  \item{visit/form/question}: Contains visit, form and question. It is
separated by a slash. This function will separate this column into `visit`,
`form` and `question` column in the output dataset. The `question` column
is the unique variable name for a particular CRF form.
  \item{description}: description of the variable, namely variable label.
This is will be used by \code{\link{apply_macro_dict}} to create variable label.
 \item{type}: type of the variable, it has IntegerData, Text, Date, Real and
Category four categories. This will be used by \code{\link{apply_macro_dict}} to
convert variables to corresponding type.
}
}
\seealso{
\code{\link{apply_macro_dict}} \code{\link{read_data}} \code{\link{cttab}}
\code{\link{report_missing}} \code{\link{clean_names}}
}
