% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{var_lab}
\alias{var_lab}
\alias{var_lab.data.frame}
\alias{var_lab<-}
\alias{has.label}
\alias{drop_lab}
\title{Set or get variable label}
\usage{
var_lab(x)

\method{var_lab}{data.frame}(x)

var_lab(x) <- value

has.label(x)

drop_lab(x)
}
\arguments{
\item{x}{Variable. In the most cases it is numeric vector.}

\item{value}{A character scalar - label for the variable x.}
}
\value{
\code{var_lab} return variable label. If label doesn't exist it return
  NULL . \code{var_lab<-} return variable (vector x) with attribute "label" which equals submitted value.
}
\description{
These functions set/get/drop variable labels. For
value labels see \link{val_lab}.
\itemize{
\item{\code{var_lab}}{ returns variable label or NULL if label doesn't
exist.}
\item{\code{var_lab<-}}{ set variable label.}
\item{\code{drop_lab}}{ drops variable label.}
\item{\code{has.label}}{ check if variable label exists.}
}
}
\details{
Variable label is stored in attribute "label" (\code{attr(x,"label")}).
To drop variable label use \code{var_lab(var) <- NULL} or \code{drop_lab(var)}.
}
\examples{
data(mtcars)
mtcars = within(mtcars,{
                var_lab(mpg) = "Miles/(US) gallon"
                var_lab(cyl) = "Number of cylinders"
                var_lab(disp) = "Displacement (cu.in.)"
                var_lab(hp) = "Gross horsepower"
                var_lab(drat) = "Rear axle ratio"
                var_lab(wt) = "Weight (lb/1000)"
                var_lab(qsec) = "1/4 mile time"
                var_lab(vs) = "V/S"
                val_lab(vs) = c("V-shaped" = 0, "straight"=1)
                var_lab(am) = "Transmission"
                val_lab(am) = c(automatic = 0, manual=1)
                var_lab(gear) = "Number of forward gears"
                var_lab(carb) = "Number of carburetors"
})

table(mtcars$am)


}
\references{
This is a modified version from `expss` package.
}
