% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{function to provide the environment in which an object is located}
\usage{
where(name, env = parent.frame())
}
\arguments{
\item{name}{character string the object name to be located}

\item{env}{the environment in which to search.}
}
\value{
an environment if the object is found, or NULL
}
\description{
function to provide the environment in which an object is located
}
\details{
the search works upwards from the environment specified in env.
}
\keyword{internal}
