% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_docx.R
\name{write_docx}
\alias{write_docx}
\title{Produces the final Docx file}
\usage{
write_docx(
  report_title,
  author,
  meta_table = get_meta_table(),
  filename = file.path(cctu_opt("output"), "Reports", "Report.docx"),
  table_path = file.path(cctu_opt("output"), "Core"),
  figure_format = "png",
  figure_path = file.path(cctu_opt("output"), "Figures"),
  popn_labels = NULL,
  verbose = options()$verbose,
  keep_xml = FALSE
)
}
\arguments{
\item{report_title}{text string used to label the report title page}

\item{author}{text string naming the author}

\item{meta_table}{a data frame that contains meta information on tables
(title, population, number). Defaults to \code{get_meta_table()}.}

\item{filename}{text string giving the filename/path to output the word
document to.}

\item{table_path}{text string giving the filepath to the tables folder.
This is used to directly open table files and copy the content.}

\item{figure_format}{it only supports \code{png} format.}

\item{figure_path}{text string giving the filepath to the figures folder.}

\item{popn_labels}{alternative text string giving labels used for the
population - might want to include the population size. They must match
\code{unique(meta_table$population)}, excluding rows with a blank population.}

\item{verbose}{logical to print information on changes to the global
environment or external files. Defaults to options()$verbose.}

\item{keep_xml}{a boolean if the compiled XML should be kept, used for
debugging purposes.}
}
\value{
This function is run for its side-effects: creates an xml document
that glues together all the outputs and meta data as per the meta-table
argument; a transformation fo this as per the xslt file, the default can
 be opened as a word document.
}
\description{
Produces the final Docx file
}
\details{
suggest that \code{\link{file.path}} is used to create non default
file paths, to cope with OS vaguaries.
}
