% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ggplot.R
\name{write_ggplot}
\alias{write_ggplot}
\title{Funtion to save ggplot figures .}
\usage{
write_ggplot(number, plot = last_plot(), width = 29.7 * 0.6, height = 21 *
  0.6, dpi = 300, units = "cm", clean_up = TRUE,
  directory = "/Output/Figures", format = c("png", "postscript", "jpg"),
  ...)
}
\arguments{
\item{number}{the figure number used to link to the meta data file}

\item{plot}{the plot object to save. defaults to \code{last_plot}}

\item{width}{the width to save as}

\item{height}{the height to save as}

\item{dpi}{the resolution setting}

\item{units}{either "cm" (the default) or "inches"}

\item{clean_up}{logical to invoke the \code{\link{clean_up}} function at the end. Defaults to TRUE}

\item{directory}{where to save the figures}

\item{format}{either "jpg", "postscript", or "png" to determine the file type to use}

\item{...}{arguments to feed into the printing function}
}
\value{
writes a copy of a plot to file fig_number.. No return object.
}
\description{
Funtion to save ggplot figures .
}
