% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_quarto.R
\name{write_quarto}
\alias{write_quarto}
\title{Produces the final HTML, PDF or DOCX file}
\usage{
write_quarto(
  report_title,
  author,
  meta_table = get_meta_table(),
  filename = file.path(getOption("cctu_output", default = "Output"), "Reports",
    "Report.html"),
  table_path = file.path(getOption("cctu_output", default = "Output"), "Core"),
  figure_path = file.path(getOption("cctu_output", default = "Output"), "Figures"),
  output_format = c("html", "pdf", "docx"),
  popn_labels = NULL,
  verbose = options()$verbose,
  quiet = TRUE
)
}
\arguments{
\item{report_title}{text string used to label the report title page}

\item{author}{text string naming the author}

\item{meta_table}{a data frame  that contains meta information on tables
(title, population, number). Defaults is get_meta_table()}

\item{filename}{text string giving the filename/path to output the word
document to}

\item{table_path}{text string giving the filepath to the tables folder.
This is used to directly open table files and copy the content.}

\item{figure_path}{text string giving the filepath to the figures folder.
This is used to create a reference within the  word document. Hence it needs
 to be a relative (to the output report) path "../Figures". Likely to lead
  to empty figures in the report if it is changed.}

\item{output_format}{Output format of the document, can be \code{"html"},
\code{"pdf"} or \code{"docx"}.}

\item{popn_labels}{alternative text string giving labels used for the
population - might want to include the population size... They must match
 correctly to unique(meta_table$population), excluding rows with a blank,
 or no, population given}

\item{verbose}{logical to print information on changes to the global
environment or external files. Defaults to options()$verbose.}

\item{quiet}{Suppress warning and other messages from quarto,
see \code{\link[quarto]{quarto_render}}}
}
\value{
This function is run for its side-effects: creates an HTML document.
}
\description{
This is for generating other document format type with Quarto. The created
\code{docx} and \code{pdf} documents may not have a nice output. Hence, the
\code{\link{write_docx}} should be used for the \code{docx} or \code{pdf}
by default. The PDF can be converted with Word. The rendering time for the
the HTML is much faster than PDF. Tables will be formated
 with \code{\link[flextable]{flextable}} and rendered with predifined Quarto
 templates using \code{\link[quarto]{quarto_render}}.
}
\details{
suggest that \code{\link{file.path}} is used to create non default
file paths, to cope with OS vaguaries.
}
