% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_text.R
\name{write_text}
\alias{write_text}
\title{Function to write a text into xml format in the correct directory, and edit
TableofTables}
\usage{
write_text(
  x,
  number = cctu_env$number,
  clean_up = TRUE,
  directory = file.path(cctu_opt("output"), "Core"),
  verbose = options()$verbose
)
}
\arguments{
\item{x}{the character string (or vector of strings) to be saved in xml
format. A vector is concatenated into a single text block.}

\item{number}{the number used as a suffix in the output filename, and to
link to the meta_table. Default is to use the value in the cctu_env
package environment that is set within \code{\link{attach_pop}}.}

\item{clean_up}{logical to invoke the \code{\link{clean_up}} function at
the end. Defaults to \code{TRUE}.}

\item{directory}{where to save the text within path or current working
directory. Defaults to \code{file.path(cctu_opt("output"), "Core")}.}

\item{verbose}{logical to print information on changes to the global
environment or external files. Defaults to \code{options()$verbose}.}
}
\value{
writes an xml version of the input data to file text_number.xml .
Edits the TableofTables object with the calling programe. No return object.
}
\description{
Function to write a text into xml format in the correct directory, and edit
TableofTables
}
\seealso{
\code{\link{write_ggplot}} \code{\link{write_table}}
}
