% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_program.R
\name{add_program}
\alias{add_program}
\title{program to add details of a code file path to the meta-table}
\usage{
add_program(number, calling_prog, meta_table_string = "meta_table", ...)
}
\arguments{
\item{number}{the number of the table or figure}

\item{calling_prog}{the file path}

\item{...}{modifications to the default values of meta_table_string, reserved_string, popn_table_string see \code{\link{clean_up}}}
}
\description{
program to add details of a code file path to the meta-table
}
\keyword{internal}
