% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_pop.R
\name{attach_pop}
\alias{attach_pop}
\alias{detach_pop}
\title{Functions that attach and detach environments based on which populations are used for a given table number}
\usage{
attach_pop(number, frame = parent.frame(), ...)

detach_pop(number, frame = parent.frame(), ...)
}
\arguments{
\item{number}{the number of a table or figure}

\item{frame}{the environment or frame in which to attach or detach the desired target environmen}

\item{...}{modifications to the default values of meta_table_string, reserved_string, popn_table_string see \code{\link{clean_up}}}

\item{meta_table_string}{character string of the name of a global table that contains meta information on tables (title, population, number)}

\item{popn_table_string}{character string of the name of a gloabl table that contains population names used in meta-table, and the equivalent names of environments}
}
\value{
invisibly returns an environment for attaching, or NULL for detaching.
}
\description{
Functions that attach and detach environments based on which populations are used for a given table number
}
\section{Functions}{
\itemize{
\item \code{attach_pop}: attaches a population

\item \code{detach_pop}: detaches a populations
}}

\seealso{
\code{\link{attach}} \code{\link{detach}}
}
