% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_pop.R
\name{attach_pop}
\alias{attach_pop}
\alias{detach_pop}
\title{Functions that attach and detach environments based on which populations are used for a given table number}
\usage{
attach_pop(number, frame = parent.frame(), verbose = options()$verbose)

detach_pop(number, frame = parent.frame(), verbose = options()$verbose)
}
\arguments{
\item{number}{a character string, or number, giving the number of a table or figure. Be careful if you want '2.10' rather than 2.1, say .}

\item{frame}{the environment or frame in which to attach or detach the desired target environment}

\item{verbose}{logical to print information on changes to the global environment or external files. Defaults to options()$verbose.}
}
\value{
invisibly returns an environment for attaching, or NULL for detaching.
}
\description{
Functions that attach and detach environments based on which populations are used for a given table number
}
\section{Functions}{
\itemize{
\item \code{attach_pop}: attaches a population

\item \code{detach_pop}: detaches a populations
}}

\seealso{
\code{\link{attach}} \code{\link{detach}}
}
