% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{init_cctu_env}
\alias{init_cctu_env}
\alias{cctu_options}
\title{Initialise cctu package environment and options}
\usage{
init_cctu_env()

cctu_options(...)
}
\arguments{
\item{...}{any options can be defined, using \code{name = value}, for example
\code{cctu_options(digits = 3)}.}
}
\description{
Allows to set and examine the global options which will affect the behaviors
of the \code{cctu} package.
}
\details{
Invoking \code{cctu_options()} with no arguments returns a list with the
current values of the options. Note that not all options listed below are set
 initially. To access the value of a single option, one should use, e.g.,
 \code{print(cctu_options("digits"))}. You need to use \code{print} in order
 to see the value.
}
\section{Options used in \code{cctu} package}{


\describe{
  \item{\code{print_plot}}{Logical, used by \code{\link{cttab}} function, if the summary
  plot will be printed.}
  \item{\code{rounding_fn}}{Name of the rounding function to used by \code{\link{cttab}}
   function. Default is \code{\link{signif_pad}}. See \code{\link{cttab}} for details.}
  \item{\code{digits_pct}}{A numeric value specifying the number of digits after
  the decimal place for percentages, default is \code{1}. Used by \code{\link{cttab}}
  function. See \code{\link{cttab}} for details.}
  \item{\code{digits}}{A numeric value of rounding to be passed to \link{rounding_fn},
   default is \code{3}. Mainly used by \code{\link{cttab}} function. See \code{\link{cttab}}
   for details.}
  \item{\code{subjid_string}}{A character string indicating the variable name of
  the subject ID in the dataset, default is "subjid"}

}
}

\examples{
\dontrun{
# No digits after decmial for percentage
cctu_options("digits_pct" = 0)
# Initialise the cctu settings.
init_cctu_env()
}

}
