% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_word_xml.R
\name{create_word_xml}
\alias{create_word_xml}
\title{Produces the final xml file}
\usage{
create_word_xml(report_title, author, meta_table,
  datestamp = format(Sys.time(), format = "\%H:\%M \%d \%b \%Y"),
  filename = "Output\\\\Reports\\\\Report.doc", path = paste0(getwd(), "/"),
  table_path = "Output\\\\Core\\\\", figure_format = c("png", "jpeg", "ps"),
  figure_path = "Output\\\\Figures\\\\", frame = parent.frame(),
  popn_labels = NULL)
}
\arguments{
\item{report_title}{text string used to label the report title page}

\item{author}{text string naming the author}

\item{meta_table}{a data frame  that contains meta information on tables (title, population, number)}

\item{datestamp}{text used to give the date-time stampe, defaults to the system date/time at the time of running the function}

\item{filename}{text string giving the filename/path to output the word document to}

\item{path}{text string giving the parent folder in which the output file, table_path and figure_path are held. Defaults to the current working directory.}

\item{table_path}{text string giving the filepath to the tables}

\item{figure_path}{text string giving the filepath to the figures}

\item{frame}{the frame or environment in which lives the meta_table to be edited with teh path to the containing code file}

\item{popn_labels}{alternative text string giving labels used for the population - might want to include the population size... They must match correctly to unique(meta_table$population)}
}
\description{
Produces the final xml file
}
\details{
all file paths need to use "\\" as "/" will not work in windows dos
}
