% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttab.R
\name{cttab}
\alias{cttab}
\title{Generate an table of descriptive statistics.}
\usage{
cttab(
  vars,
  data,
  group = NULL,
  row_split = NULL,
  total = TRUE,
  select = NULL,
  add_missing = TRUE,
  add_obs = TRUE,
  digits = getOption("cctu_digits", default = 3),
  digits_pct = getOption("cctu_digits_pct", default = 0),
  rounding_fn = signif_pad,
  dlu = cctu_env$dlu,
  subjid_string = getOption("cctu_subjid_string", default = "subjid"),
  print_plot = getOption("cctu_print_plot", default = TRUE)
)
}
\arguments{
\item{vars}{Variables to be used for summary table.}

\item{data}{A \code{data.frame} from which the variables in \code{vars}
should be taken.}

\item{group}{Name of the grouping variable.}

\item{row_split}{Variable that used for splitting table rows, rows will be
splited using this variable. Useful for repeated measures.}

\item{total}{If a "Total" column will be created (default). Specify
\code{FALSE} to omit the column.}

\item{select}{a named vector with as many components as row-variables. Every
element of `select` will be used to select the individuals to be analyzed
 for every row-variable. Name of the vector corresponds to the row variable,
 element is the selection.}

\item{add_missing}{If missing number and missing percentage will be
reported in the summary table, default is `TRUE`.}

\item{add_obs}{Add an observation row (default).}

\item{digits}{An integer specifying the number of significant digits to keep,
default is 3.}

\item{digits_pct}{An integer specifying the number of digits after the
decimal place for percentages, default is 0.}

\item{rounding_fn}{The function to use to do the rounding. Defaults is
\code{\link{signif_pad}}. To round up by digits instead of significant
values, set it to \code{round_pad}.}

\item{dlu}{A data.frame of DLU file.}

\item{subjid_string}{A character naming the column used to identify subject,
default.}

\item{print_plot}{A logical value, print summary plot of the variables (default).}
}
\value{
A matrix with `cttab` class.
}
\description{
This is a wrapper function of \code{stat_tab}, allowing for groupped variables,
split statistics table by `row_split` variable.
}
\details{
Some of the function parameters can be set with options. This will have an global
effect on the \code{cctab} function. It is an ideal way to set a global settings
if you want this to be effctive globally. Currently, you can set \code{digits},
\code{digits_pct}, \code{subjid_string} and \code{print_plot}  by adding \code{"cctu_"}
 prefix in the \code{options}. For example, you can suppress the plot
from printting by setting \code{options(cctu_print_plot = FALSE)}.
}
\seealso{
\code{\link{signif_pad}}
\code{\link{round_pad}}
\code{\link{stat_tab}}
\code{\link{sumby}}
\code{\link{dump_missing_report}}
\code{\link{get_missing_report}}
}
