% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttab.R
\name{cttab}
\alias{cttab}
\alias{cttab.default}
\alias{cttab.formula}
\title{Generate an table of descriptive statistics.}
\usage{
cttab(x, ...)

\method{cttab}{default}(
  x,
  data,
  group = NULL,
  row_split = NULL,
  total = TRUE,
  select = NULL,
  add_missing = TRUE,
  add_obs = TRUE,
  digits = getOption("cctu_digits", default = 3),
  digits_pct = getOption("cctu_digits_pct", default = 0),
  rounding_fn = signif_pad,
  subjid_string = getOption("cctu_subjid_string", default = "subjid"),
  print_plot = getOption("cctu_print_plot", default = TRUE),
  render_num = getOption("cctu_render_num", default = "Median [Min, Max]"),
  logical_na_impute = c(FALSE, NA, TRUE),
  blinded = getOption("cctu_blinded", default = FALSE),
  ...
)

\method{cttab}{formula}(
  x,
  data,
  total = TRUE,
  select = NULL,
  add_missing = TRUE,
  add_obs = TRUE,
  digits = getOption("cctu_digits", default = 3),
  digits_pct = getOption("cctu_digits_pct", default = 0),
  rounding_fn = signif_pad,
  subjid_string = getOption("cctu_subjid_string", default = "subjid"),
  print_plot = getOption("cctu_print_plot", default = TRUE),
  render_num = getOption("cctu_render_num", default = "Median [Min, Max]"),
  logical_na_impute = c(FALSE, NA, TRUE),
  blinded = getOption("cctu_blinded", default = FALSE),
  ...
)
}
\arguments{
\item{x}{Variables to be used or a \code{formula} for summary table.
If \code{x} is a \code{formula}, then the \code{group} variable should
 be provided at the right had side, use \code{1} if there's no grouping
variable. And \code{row_split} should also be provided on the right hand side
of the formula and separate it using \code{|} with grouping variable. For example,
\code{age + sex ~ treat|cycle} or \code{age + sex ~ 1|cycle} without grouping.
See details.}

\item{...}{Not used.}

\item{data}{A \code{data.frame} from which the variables in \code{vars}
should be taken.}

\item{group}{Name of the grouping variable.}

\item{row_split}{Variable that used for splitting table rows, rows will be
splited using this variable. Useful for repeated measures.}

\item{total}{If a "Total" column will be created (default). Specify
\code{FALSE} to omit the column.}

\item{select}{a named vector with as many components as row-variables. Every
element of `select` will be used to select the individuals to be analyzed
 for every row-variable. Name of the vector corresponds to the row variable,
 element is the selection.}

\item{add_missing}{If missing number and missing percentage will be
  reported in the summary table, default is `TRUE`. This will also produce
data missingness report if set \code{TRUE}. See \code{\link{report_missing}} 
for details.}

\item{add_obs}{Add an observation row (default).}

\item{digits}{An integer specifying the number of significant digits to keep,
default is 3.}

\item{digits_pct}{An integer specifying the number of digits after the
decimal place for percentages, default is 0.}

\item{rounding_fn}{The function to use to do the rounding. Defaults is
\code{\link{signif_pad}}. To round up by digits instead of significant
values, set it to \code{round_pad}.}

\item{subjid_string}{A character naming the column used to identify subject,
default is \code{"subjid"}.}

\item{print_plot}{A logical value, print summary plot of the variables (default).}

\item{render_num}{A character or vector indicatin which summary will be reported,
 default is "Median [Min, Max]". You can change this to "Median [IQR]" then the
median and IQR will be reported instead of "Median [Min, Max]". Use
\code{options(cctu_render_num = "Median [IQR]")} to set global options.
See details \code{\link{render_numeric}}.}

\item{logical_na_impute}{Impute missing values with \code{FALSE} (default),
\code{NA} keep as it is, or \code{TRUE}. The nominator for the logical vector is
the number of \code{TRUE}. For \code{FALSE} or \code{TRUE}, the denominator will
 be all values regardless of missingness, but the non-missing number used as
denominator for \code{NA}. Set it to \code{FALSE} if you want to summarise multiple
choice variables and \code{NA} for Yes/No type logical variables but don't want No
in the summary. You can used a named list in \code{x} and stack multiple
 choice in one category.}

\item{blinded}{A logical scalar, if summary table will be report by
\code{group} (default) or not. This will ignore \code{group} if set to \code{TRUE}
and grouping summary will not be reported.}
}
\value{
A matrix with `cttab` class.
}
\description{
This is a wrapper function of \code{stat_tab}, allowing for groupped variables,
split statistics table by `row_split` variable.
}
\details{
\strong{1. Parameter settings with global options}

Some of the function parameters can be set with options. This will have an global
effect on the \code{cctab} function. It is an ideal way to set a global settings
if you want this to be effctive globally. Currently, you can set \code{digits},
\code{digits_pct}, \code{subjid_string}, \code{print_plot}, \code{render_num} and
\code{blinded}  by adding \code{"cctu_"} prefix in the \code{options}. For example,
 you can suppress the plot from printting by setting \code{options(cctu_print_plot = FALSE)}.

\strong{2. Formula interface}

There are two interfaces, the default, which typically takes a variable vector from
\code{data.frame} for \code{x}, and the formula interface. The formula interface is
 less flexible, but simpler to use and designed to handle the most common use cases.
For the formula version, the formula is expected to be a two-sided formula. Left hand
side is the variables to be summarised and the right hand side is the group and/or split
 variable. To include a row splitting variable, use \code{|} to separate the row splitting
variable after the grouping variable and then the row split variable. For example,
\code{age + sex ~ treat|visit}. The right hand side of the formula will be treated as a grouping
 variable by default. A value of \code{1} should be provided if there is no grouping variable,
 for example \code{age + sex ~ 1} or \code{age + sex ~ 1|visit} by visit.

\strong{3. Return}

A summary table with some attributes will be reutned, a method has been writen for \code{rbind}.
So you can use \code{rbind} to combine two tables without losing any attributes. An attribute
\code{position} will be used to produce a nice table. There are three 4 possible values for each
rows. Row name printed as the first column in the word table. Some styles will be applied to each
row based on the \code{position} attributes.
\tabular{ll}{
\code{0} \tab indicates the row will be bolded, spanned through all columns and a grey background
in the word \cr
\tab \cr
\code{1} \tab indicates the row will be bolded \cr
\tab \cr
\code{2} \tab the row will be bolded and spanned through all columns \cr
\tab \cr
\code{3} \tab indicates the row of the first column will be indented \cr
}
}
\section{Methods (by class)}{
\itemize{
\item \code{cttab(default)}: The default interface, where \code{x} is a \code{data.frame}.

\item \code{cttab(formula)}: The formula interface, where \code{x} is a \code{formula}.

}}
\examples{

# Read data
dt <- read.csv(system.file("extdata", "pilotdata.csv", package="cctu"))
dlu <- read.csv(system.file("extdata", "pilotdata_dlu.csv", package="cctu"))
clu <- read.csv(system.file("extdata", "pilotdata_clu.csv", package="cctu"))

dt$subjid <- substr(dt$USUBJID, 8, 11)

# Apply variable attributes
dt <- apply_macro_dict(dt, dlu, clu, clean_names = FALSE)

# Extract form data to be analysed
df <- extract_form(dt, "PatientReg", vars_keep = c("subjid"))

########################################################
#  Simple analysis no group and variable subset
######################################################
# Variable as a vector
X <- cttab(x = c("AGE", "SEX", "BMIBL"),
             data = df,
             select = c("BMIBL" = "RACEN != 1"))

# Variable as a formula, equivalent to above
X1 <- cttab(AGE + SEX + BMIBL ~ 1,
            data = df,
            select = c("BMIBL" = "RACEN != 1"))

#############################################
#  Analysis by group
############################################
# Variable as a vector
X <- cttab(x = c("AGE", "SEX", "BMIBL"),
             group = "ARM",
             data = df,
             select = c("BMIBL" = "RACEN != 1"))

############################################
# Analysis by group and cycles
############################################

df <- extract_form(dt, "Lab", vars_keep = c("subjid", "ARM"))

X <- cttab(x = c("AST", "BILI", "ALT"),
                  group = "ARM",
                  data = df,
                  row_split = "AVISIT",
                  select = c("ALT" = "PERF == 1"))

############################################
# Group variables
############################################

df <- extract_form(dt, "PatientReg", vars_keep = c("subjid"))
base_lab <- extract_form(dt, "Lab", visit = "SCREENING", vars_keep = c("subjid"))

base_lab$ABNORMALT <- base_lab$ALT > 22.5
var_lab(base_lab$ABNORMALT) <- "ALT abnormal"
base_lab$ABNORMAST <- base_lab$AST > 25.5
var_lab(base_lab$ABNORMAST) <- "AST abnormal"

df <- merge(df, base_lab, by = "subjid")

X <- cttab(x = list(c("AGE", "SEX", "BMIBL"),
                      "Blood" = c("ALT", "AST"),
                      "Patients with Abnormal" = c("ABNORMAST", "ABNORMALT")),
          group = "ARM",
          data = df,
          select = c("BMIBL" = "RACEN != 1",
                     "ALT" = "PERF == 1"))

}
\seealso{
\code{\link{signif_pad}}
\code{\link{round_pad}}
\code{\link{stat_tab}}
\code{\link{sumby}}
\code{\link{dump_missing_report}}
\code{\link{get_missing_report}}
\code{\link{render_numeric}}
\code{\link{render_cat}}
}
