% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttab.R
\name{cttab}
\alias{cttab}
\title{Generate an table of descriptive statistics.}
\usage{
cttab(
  vars,
  data,
  group = NULL,
  row_split = NULL,
  total = TRUE,
  select = NULL,
  add_missing = TRUE,
  add_obs = TRUE,
  digits = cctu_options("digits"),
  digits_pct = cctu_options("digits_pct"),
  rounding_fn = cctu_options("rounding_fn"),
  dlu = cctu_options("dlu"),
  subjid_string = cctu_options("subjid_string"),
  print_plot = cctu_options("print_plot")
)
}
\arguments{
\item{vars}{Variables to be used for summary table.}

\item{data}{A \code{data.frame} from which the variables in \code{vars}
should be taken.}

\item{group}{Name of the grouping variable.}

\item{row_split}{Variable that used for splitting table rows, rows will be
splited using this variable. Useful for repeated measures.}

\item{total}{If a "Total" column will be created (default). Specify
\code{FALSE} to omit the column.}

\item{select}{a named vector with as many components as row-variables. Every
element of `select` will be used to select the individuals to be analyzed
 for every row-variable. Name of the vector corresponds to the row variable,
 element is the selection.}

\item{add_missing}{If missing number and missing percentage will be
reported in the summary table, default is `TRUE`.}

\item{add_obs}{Add an observation row (default).}

\item{digits}{An integer specifying the number of significant digits to keep.}

\item{digits_pct}{An integer specifying the number of digits after the
decimal place for percentages.}

\item{rounding_fn}{The function to use to do the rounding. Defaults to
\code{\link{signif_pad}}. To round up by digits instead of significant
values, set it to \code{round_pad}.}

\item{dlu}{A data.frame of DLU file.}

\item{subjid_string}{A character naming the column used to identify subject.}

\item{print_plot}{A logical value, print summary plot of the variables (default).}
}
\value{
A matrix with `cttab` class.
}
\description{
This is a wrapper function of \code{stat_tab}, allowing for groupped variables,
split statistics table by `row_split` variable.
}
\seealso{
\code{\link{signif_pad}}
\code{\link{round_pad}}
\code{\link{stat_tab}}
\code{\link{sumby}}
\code{\link{cctu_options}}
}
