% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_utils.R
\name{extract_form}
\alias{extract_form}
\title{Extract data by form from MACRO dataset}
\usage{
extract_form(data, form, visit = NULL, vars_keep = NULL, dlu = cctu_env$dlu)
}
\arguments{
\item{data}{A data.frame from macro dataset.}

\item{form}{Name of the form in the DLU file, see \code{\link{sep_dlu}}.}

\item{visit}{A character string or vector of visit name in the DLU file, see \code{\link{sep_dlu}}.}

\item{vars_keep}{Parameters to keep in the output data. This is useful if you want to keep
treatment arm or age variable.}

\item{dlu}{A DLU data.frame}
}
\value{
A data.table object.
}
\description{
Extract data by form from MACRO dataset. Data will be transformed to long format
 adding a new column of `Visit`.
}
\examples{
\dontrun{
extract_form(full_dt, "LabResF", dlu)
}

}
\seealso{
\code{\link{sep_dlu}} \code{\link[data.table]{data.table}}
}
