% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_table_functions.R
\name{get_meta_table}
\alias{get_meta_table}
\alias{set_meta_table}
\alias{add_program}
\title{Functions to handle meta_table}
\usage{
get_meta_table()

set_meta_table(meta_table)

add_program(number, calling_prog)
}
\arguments{
\item{meta_table}{a data.frame to be  set as the meta_table internal object}

\item{number}{the number of the table or figure}

\item{calling_prog}{the file path}
}
\value{
get_table returns the global object, set_table invisibly returns the previous version
}
\description{
Functions to handle meta_table
}
\section{Functions}{
\itemize{
\item \code{get_meta_table}: gets a copy of the internal object

\item \code{set_meta_table}: sets the internal object to the argument provided, whilst carrying out some cleaning

\item \code{add_program}: edits the internal meta_table to gve the filepath to code that created a table/figure
}}

\keyword{internal}
