% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_blank_rows_cols.R
\name{remove_blank_rows_cols}
\alias{remove_blank_rows_cols}
\title{Function to remove blank rows/columns from a df}
\usage{
remove_blank_rows_cols(
  df,
  convert = TRUE,
  verbose = options()$verbose,
  which = c("both", "rows", "cols")
)
}
\arguments{
\item{df}{a data.frame}

\item{convert}{a logical to indicate if you want to modify \code{df} in the
parent environment, or if not simply return a modified version of \code{df}}

\item{verbose}{logical to print information on changes to the global
environment or external files. Defaults to options()$verbose.}

\item{which}{one of \code{"rows"}, \code{"cols"}, or \code{"both"}. Default
is \code{"both"} to removing both empty rows and empty columns.}
}
\value{
this removes columns and rows that are totally
empty (either "" or NA).
}
\description{
Function to remove blank rows/columns from a df
}
