% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_stat.R
\name{render_numeric}
\alias{render_numeric}
\title{Render continuous values for table output.}
\usage{
render_numeric(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Further arguments, passed to \code{\link{num_stat}}.}
}
\value{
A \code{character} vector. Each element is to be displayed in a
separate cell in the table. The \code{\link{names}} of the vector are the
labels to use in the table.
}
\description{
Called from \code{\link{cttab}} by default to render continuous (i.e.
\code{numeric}) values for displaying in the table.
}
\examples{
x <- exp(rnorm(100, 1, 1))
render_numeric(x)

}
\keyword{utilities}
