% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_envir.R
\name{rm_envir}
\alias{rm_envir}
\title{A function to remove all attached objects from the search path with any exceptions specified}
\usage{
rm_envir(
  ignore = c(".GlobalEnv", "package:", "tools:", "Autoloads"),
  verbose = options()$verbose,
  perl = FALSE
)
}
\arguments{
\item{ignore}{a character vector of regular expression search terms  within the output from \code{search()} as to which should be retained.}

\item{verbose}{logical to print information on changes to the global environment or external files. Defaults to options()$verbose.}

\item{perl}{logical.  Should Perl-compatible regexps be used?}
}
\description{
A function to remove all attached objects from the search path with any exceptions specified
}
\details{
it detaches anything not ignored. It is called by default wihin \code{\link{attach_pop}} .
}
\seealso{
\code{\link{search}} \code{\link{attach_pop}}
}
