% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_utils.R
\name{sep_dlu}
\alias{sep_dlu}
\alias{set_dlu}
\alias{get_dlu}
\title{Tidy DLU form}
\usage{
sep_dlu(x)

set_dlu(x)

get_dlu()
}
\arguments{
\item{x}{DLU data}
}
\value{
A data.frame
}
\description{
Separate visit, form and question into different columns.
The DLU file contains four columns:
\itemize{
  \item{shortcode}: Variable name of the downloaded data.
  \item{visit/form/question}: Contains visit, form and question. It is
separated by a slash. This function will separate this column into `visit`,
`form` and `question` column in the output dataset. The `question` column
is the unique variable name for a particular CRF form.
  \item{description}: description of the variable, namely variable label.
This is will be used by \code{\link{apply_macro_dict}} to create variable label.
 \item{type}: type of the variable, it has IntegerData, Text, Date, Real and
Category four categories. This will be used by \code{\link{apply_macro_dict}} to
convert variables to corresponding type.
}
Variable names, NOT the values, of the dlu data will be converted to lower cases.

\code{set_dlu} will set the provided DLU data to package environment,
 so it can be used for missing data report by \code{\link{cttab}}. It is user's
 responsibility to make sure values of the short code in the provided DLU data matches
 the variable names in the dataset.

\code{set_dlu} can be used to get a copy of DLU data in stored by \code{set_dlu}.
}
\seealso{
\code{\link{apply_macro_dict}} \code{\link{read_data}} \code{\link{cttab}}
\code{\link{report_missing}}
}
