% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_utils.R
\name{sep_dlu}
\alias{sep_dlu}
\title{Separates out visit/form/question from DLU}
\usage{
sep_dlu(x, clean_names = TRUE)
}
\arguments{
\item{x}{DLU data}

\item{clean_names}{Conver variable name to lower case (default). See
\code{\link{clean_names}} for details.}
}
\value{
A data.frame with expanded, separated variables
}
\description{
Deprecated so included for back compatibility
}
\seealso{
\code{\link{tidy_dlu}}
}
