% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source}
\alias{source}
\title{A modified version of the source() function to capture in a data.frame what
file was called and from which file}
\usage{
source(
  file,
  local = getOption("cctu_source_local", default = FALSE),
  backup = NULL,
  ...
)
}
\arguments{
\item{file}{a \link[base]{connection} or a character string giving the
    pathname of the file or URL to read from.  The \code{\link[base]{stdin}()}
    connection reads from the console when interactive.}

\item{local}{\code{TRUE}, \code{FALSE} or an environment, determining
    where the parsed expressions are evaluated.  \code{FALSE} (the
    default) corresponds to the user's workspace (the global
    environment) and \code{TRUE} to the environment from which
    \code{source} is called.}

\item{backup}{A character string giving the file path to which to save an
image of the R environment before sourcing any code.
Default is NULL, which does not save any image.}

\item{...}{other parameters to be passed to \code{\link[base]{source}}.}
}
\value{
no return value
}
\description{
A modified version of the source() function to capture in a data.frame what
file was called and from which file
}
\details{
evaluates code in the file argument, and also augments or creates a
 dataframe with two variables (parent, child). The intention is to record
 the architecture of a sequenece of code files run using nested source()
 statements. To override this function use
\code{source <- base::source}.
}
\seealso{
\code{\link[base]{source}}
}
