% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srce.R
\name{srce}
\alias{srce}
\title{Wrapper function for \code{\link{source}}}
\usage{
srce(file, backup = FALSE, path_string = "PATH",
  backup_filepath = "\\\\Output\\\\R_images\\\\backup.Rdata")
}
\arguments{
\item{file}{a file to source}

\item{backup}{a logical to indicate if you want to save an image prior to sourcing the course}

\item{path_string}{character string of the name of a global variable that contains the project filepath. Default is "PATH".}

\item{backup_filepath}{a filepath to where to save the image}
}
\description{
Wrapper function for \code{\link{source}} to save typing/avoid errors, override some defaults, and save a backup copy
}
