% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttab.R
\name{stat_tab}
\alias{stat_tab}
\title{Generate a descriptive summary statistics table.}
\usage{
stat_tab(
  vars,
  group = NULL,
  data,
  total = TRUE,
  select = NULL,
  add_missing = TRUE,
  digits = 2,
  digits_pct = 1,
  rounding_fn = signif_pad
)
}
\arguments{
\item{group}{Name of the grouping variable.}

\item{data}{A \code{data.frame} from which the variables in \code{vars}
should be taken.}

\item{total}{If a "Total" column will be created (default). Specify
\code{FALSE} to omit the column.}

\item{select}{a named vector with as many components as row-variables. Every
element of `select` will be used to select the individuals to be analyzed
 for every row-variable. Name of the vector corresponds to the row variable,
 element is the selection.}

\item{add_missing}{If missing number and missing percentage will be
  reported in the summary table, default is `TRUE`. This will also produce
data missingness report if set \code{TRUE}. See \code{\link{report_missing}} 
for details.}

\item{digits}{An integer specifying the number of significant digits to keep,
default is 3.}

\item{digits_pct}{An integer specifying the number of digits after the
decimal place for percentages, default is 0.}

\item{rounding_fn}{The function to use to do the rounding. Defaults is
\code{\link{signif_pad}}. To round up by digits instead of significant
values, set it to \code{round_pad}.}
}
\value{
An object of class "cttab".
}
\description{
It is important to use variable label and value label to produce a proper
descriptive table. Variables with value labels will be converted to ordered
factor with same order as the value labels (\code{to_factor}). And variable
labels will be used in the output. The first row will be blank with row names
of variable label. Variable name will be used if the variable does not have
a variable label.
}
\keyword{internal}
