% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttab.R
\name{stat_tab}
\alias{stat_tab}
\title{Generate a descriptive summary statistics table.}
\usage{
stat_tab(
  vars,
  group = NULL,
  data,
  total = TRUE,
  select = NULL,
  add_missing = TRUE,
  digits = 2,
  digits_pct = 1,
  rounding_fn = signif_pad
)
}
\arguments{
\item{vars}{Variables to be used for summary table.}

\item{group}{Name of the grouping variable.}

\item{data}{A \code{data.frame} from which the variables in \code{vars}
should be taken.}

\item{total}{If a "Total" column will be created (default). Specify
\code{FALSE} to omit the column.}

\item{select}{a named vector with as many components as row-variables. Every
element of `select` will be used to select the individuals to be analyzed
 for every row-variable. Name of the vector corresponds to the row variable,
 element is the selection.}

\item{add_missing}{If missing number and missing percentage will be
reported in the summary table, default is `TRUE`.}

\item{digits}{An integer specifying the number of significant digits to keep.}

\item{digits_pct}{An integer specifying the number of digits after the
decimal place for percentages.}

\item{rounding_fn}{The function to use to do the rounding. Defaults to
\code{\link{signif_pad}}. To round up by digits instead of significant
values, set it to \code{round_pad}.}
}
\value{
An object of class "cttab".
}
\description{
It is important to use variable label and value label to produce a proper
descriptive table. Variables with value labels will be converted to ordered
factor with same order as the value labels (\code{to_factor}). And variable
labels will be used in the output. The first row will be blank with row names
of variable label. Variable name will be used if the variable does not have
a variable label.
}
\examples{

dat <- expand.grid(id=1:10, sex=c("Male", "Female"), treat=c("Treated", "Placebo"))
dat$age <- runif(nrow(dat), 10, 50)
dat$age[3] <- NA  # Add a missing value
dat$wt <- exp(rnorm(nrow(dat), log(70), 0.2))

var_lab(dat$sex) <- "Sex"
var_lab(dat$age) <- "Age"
var_lab(dat$treat) <- "Treatment Group"
var_lab(dat$wt) <- "Weight"


# Something more complicated

dat$dose <- ifelse(dat$treat=="Placebo", "Placebo",
                   sample(c("5 mg", "10 mg"), nrow(dat), replace=TRUE))
dat$dose <- factor(dat$dose, levels=c("Placebo", "5 mg", "10 mg"))


cttab(c("age", "sex", "wt"),
data = dat,
group = "treat")
}
\keyword{utilities}
