% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ggplot.R
\name{write_ggplot}
\alias{write_ggplot}
\title{Function to save ggplot figures.}
\usage{
write_ggplot(number, plot = last_plot(), width = 29.7 * 0.6, height = 21 *
  0.6, dpi = 300, units = "cm", clean_up = TRUE,
  directory = "Output/Figures/", path_string = "PATH", ...,
  format = c("png", "postscript", "jpeg"), graphics_args = NULL,
  frame = parent.frame())
}
\arguments{
\item{number}{the number used to as a suffix in the output filename, and to link to TableofTables.}

\item{plot}{the plot object to save. defaults to \code{last_plot}}

\item{width}{the width to save as}

\item{height}{the height to save as}

\item{dpi}{the resolution setting}

\item{units}{either "cm" (the default) or "inches"}

\item{clean_up}{logical to invoke the \code{\link{clean_up}} function at the end. Defaults to TRUE}

\item{directory}{where to save the figures}

\item{path_string}{character string of the name of a global variable that contains the project filepath. Default is "PATH".}

\item{...}{modifications to the default values of meta_table_string, reserved_string, popn_table_string see \code{\link{clean_up}}}

\item{format}{either "jpg", "postscript", or "png" to determine the file type to use}

\item{graphics_args}{a list of named arguments to supply to graphics function (png, postscript, jpeg)}
}
\value{
writes a copy of a plot to file fig_number.. edits the TableofTables object with the calling programe No return object.
}
\description{
Function to save ggplot figures.
}
\seealso{
\code{\link{get_file_name}} \code{\link{write_table}}
}
