% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ggplot.R
\name{write_ggplot}
\alias{write_ggplot}
\title{Function to save ggplot figures.}
\usage{
write_ggplot(plot = last_plot(), number = cctu_env$number,
  width = 29.7 * 0.6, height = 21 * 0.6, dpi = 300, units = "cm",
  clean_up = TRUE, directory = "Output\\\\Figures\\\\", ...,
  format = c("png", "postscript", "jpeg"), graphics_args = NULL,
  frame = parent.frame())
}
\arguments{
\item{plot}{the plot object to save. defaults to \code{last_plot}}

\item{number}{the number used to as a suffix in the output filename, and to link to TableofTables. Default is to use the value in the cctu_env package environment that is set within \code{\link{attach_pop}}.}

\item{width}{the width to save as}

\item{height}{the height to save as}

\item{dpi}{the resolution setting}

\item{units}{either "cm" (the default) or "inches"}

\item{clean_up}{logical to invoke the \code{\link{clean_up}} function at the end. Defaults to TRUE}

\item{directory}{where to save the figures within path or current working directory}

\item{...}{modifications to the default values \code{\link{detach_pop}}}

\item{format}{either "jpg", "postscript", or "png" to determine the file type to use}

\item{graphics_args}{a list of named arguments to supply to graphics function (png, postscript, jpeg)}

\item{frame}{the frame or environment in which lives the meta_table to be edited with teh path to the containing code file}
}
\value{
writes a copy of a plot to file fig_number.. edits the TableofTables object with the calling programe No return object.
}
\description{
Function to save ggplot figures.
}
\seealso{
\code{\link{get_file_name}} \code{\link{write_table}}
}
