% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{write_table}
\alias{write_table}
\title{Function to write a table into xml format in the correct directory, and edit TableofTables}
\usage{
write_table(number, X, heading = colnames(X), na_to_empty = FALSE,
  clean_up = TRUE, directory = "Output/Core/", path = NULL,
  frame = parent.frame(), ...)
}
\arguments{
\item{number}{the number used to as a suffix in the output filename, and to link to TableofTables.}

\item{X}{the data.frame or table to be saved in xml format}

\item{heading}{character vector of column titles. Defaults to the colnames of X}

\item{na_to_empty}{logical, if true then any NA values will be written as empty strings. Defaults to false.}

\item{clean_up}{logical to invoke the \code{\link{clean_up}} function at the end. Defaults to TRUE}

\item{directory}{where to save the figures within path or current working directory}

\item{path}{character string of the project filepath to save the table in. Default is NULL which implies the current working directory.}

\item{frame}{the frame or environment in which lives the meta_table to be edited with teh path to the containing code file}

\item{...}{modifications to the default values \code{\link{detach_pop}}}
}
\value{
writes an xml version of the input data to file table_number.xml . Edits the TableofTables object with the calling programe. No return object.
}
\description{
Function to write a table into xml format in the correct directory, and edit TableofTables
}
\seealso{
\code{\link{get_file_name}} \code{\link{write_ggplot}}
}
