% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{write_table}
\alias{write_table}
\title{Function to write a table into xml format in the correct directory, and edit
TableofTables}
\usage{
write_table(
  X,
  number = cctu_env$number,
  heading = colnames(X),
  na_to_empty = getOption("cctu_na_to_empty", default = FALSE),
  clean_up = TRUE,
  directory = file.path("Output", "Core"),
  verbose = options()$verbose
)
}
\arguments{
\item{X}{the data.frame or table to be saved in xml format}

\item{number}{the number used to as a suffix in the output filename, and to link to TableofTables. Default is to use the value in the cctu_env package environment that is set within \code{\link{attach_pop}}.}

\item{heading}{character vector of column titles. Defaults to the colnames of X}

\item{na_to_empty}{logical, if true then any NA values will be written as empty strings. Defaults to false.}

\item{clean_up}{logical to invoke the \code{\link{clean_up}} function at the end. Defaults to TRUE}

\item{directory}{where to save the figures within path or current working directory}

\item{verbose}{logical to print information on changes to the global environment or external files. Defaults to options()$verbose.}
}
\value{
writes an xml version of the input data to file table_number.xml . Edits the TableofTables object with the calling programe. No return object.
}
\description{
Function to write a table into xml format in the correct directory, and edit
TableofTables
}
\details{
Variable names and values will be replace by variable labels and value labels respectively if available before writing the data.
Use \code{options(cctu_na_to_empty = TRUE)} to write NA values will be written as empty strings globally.
}
\seealso{
\code{\link{write_ggplot}} \code{\link{detect_invalid_utf8}} \code{\link{remove_invalid_utf8}} 
\code{\link{lab2val}} \code{\link{var_lab}} \code{\link{val_lab}} \code{\link{apply_macro_dict}}
}
