% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_text.R
\name{write_text}
\alias{write_text}
\title{Function to write a text into xml format in the correct directory, and edit TableofTables}
\usage{
write_text(
  X,
  number = cctu_env$number,
  clean_up = TRUE,
  directory = file.path("Output", "Core"),
  verbose = options()$verbose
)
}
\arguments{
\item{X}{the character string to be saved in xml format}

\item{number}{the number used to as a suffix in the output filename, and to link to TableofTables. Default is to use the value in the cctu_env package environment that is set within \code{\link{attach_pop}}.}

\item{clean_up}{logical to invoke the \code{\link{clean_up}} function at the end. Defaults to TRUE}

\item{directory}{where to save the figures within path or current working directory}

\item{verbose}{logical to print information on changes to the global environment or external files. Defaults to options()$verbose.}
}
\value{
writes an xml version of the input data to file text_number.xml . Edits the TableofTables object with the calling programe. No return object.
}
\description{
Function to write a text into xml format in the correct directory, and edit TableofTables
}
\seealso{
\code{\link{write_ggplot}} \code{\link{write_table}}
}
